scripts/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
supabase==2.0.3
python-dotenv==1.0.0
openai==1.3.0